Adding a global functions file.
After using the CFWheels framework, I just can't stop thinking about that `e()` alias for `encodeForHtml()`. It's just so freakin' nice. I need to have it in my life. This commit starts to lay the foundation for a file of global functions that is intended to be included into any context in which it is needed / desired.

Of course, this is a risky proposition. Any "global" function runs the risk of colliding with other functions that are specific to a given template. As such, these functions are intended to be thought of as CFML extensions - extensions to the language itself; and should pertain primarily to polyfills and data structure manipulation. In other words, I think this should be limited to functions that might be added to CFML in the future.

Many of these were copied out of the `Utilities.cfc`. Once this is integrated, much of that component will be gutted.