Fix memory leak detectory for CFMLX.
Since the CFMLX functions aren't in the component metadata, they don't get seen as native to the inspected components; and, were showing up as memory leaks. This adds a hack that skips over any Function that is in both the target component and the memory leak detector (since it also includes the CFMLX functions).