Fixing flash translations.
I had a bunch of flash translations copied from an old project. The flash translations should all work now. I'm exploring the idea of using a `your.` prefix in the flash token to help differentiate future messages which are personal vs. administrative. Might not be worthwhile, but easy to change in the future.