Cache syllable count results for all Datamuse responses.
When I go to Datamuse for rhymes and synonyms, all of the responses _also_ contain the syllable count. I might as well start caching these now so that I don't have to cache them later when the users are typing them. This may turn out to be a bad idea (volume-wise); but, let's see how it goes.