Save poem in composer background.
This implements the incremental save-in-background in the composer UI. this UI is the first real use of HTMX and I'm changing some of my thinking on it. I had originally copied over the `responseHandling` configuration from previous experiments. But, in those experiments, I was using `hx-boost`; and therefore needed non `2xx` errors to be rendered. In this app, however, I'm not using `hx-boost` (using full page navigation); and therefore, I don't need to render error pages. As such, I've removed the `responseHandling` overrides.

That said, I've also created a special `processErrorForHtmx()` helper that will log the error and translate the response message but _won't_ change the status code.