Noodling on a toast module for htmx errors.
This is an incomplete thought; but, it gets me started down a path to showing htmx errors in a toast. I don't want the user to just sit there wondering if anything is happening - an error should always be rendered in some way so as to inform the user. But, the implementation isn't complete.