Removing implicit dates from model layer.
I'm on the fence about this notion, but I'm moving the implicit `createdAt`/`updatedAt` date creation out of the model and into the service layer. As much as possible, I just want the models to be "dumb". It feels strange to make this as a sweeping change; I suspect it will feel less strange when I do new model usage.