Add meta-enter submit for textarea elements.
This is a common pattern on the web. It allows the user to submit form content directly from the keyboard just as if they were focused on an input element.

This is my first Alpine.js component. I'm not entirely sure that it's wired up correctly. That is, I'm not entirely sure that the `Alpine.start()` method is always going to be called at the right time. I'll have to explore that a bit more.