Rotating validation method names.
I'm trying a new strategy. Instead of having validation methods like:

* `testName( input )`
* `testContent( input )`

... I'm rotating the tokens so that the property names comes first:

* `nameFrom( input )`
* `contentFrom( input )`

It's a little unusual sounding; but, it will make copy-paste a lot easier since I never have to change the casing of the properties (where as before, `name` would have to be re-cased as `Name`). Also, I think that `test` didn't give much indication of normalization whereas `from` is a little more clear that a _transformation_ is taking place. And the "testing" portion can be implied since the method is located in the `xxxValidation.cfc` class.