Moving to a top-nav-bar.
The site is very simple. As such, having a side-nav-bar with a differentiation between "member" and "account" sections is overkill for the time-being. I'm collapsing the two major areas and also moving to a top-nav-bar. This should also make more room for the poem composer UI, which is the kind of the main point of the site.