Update session access to use `SessionAccess` pattern.
I was giving special treatment to "sessions"; which, to be fair, still get some special treatment during creation. But, deleting a session is now treated like deleting any other domain model in that it's access permissions are checked against the currently authenticated user (instead of relying on the Controller to just pass-in the right userID).