Make IoC and CFMLX consistent.
This just satisfies my need to have some consistency. This adds the IoC "comment" and the `cfmlx.cfm` include to every module. In a number of these cases, I'm not even using any of the `cfmlx` functions. But, I want to lean into the idea that these are just everywhere. If this ever turns out to be a performance problem, I can deal with it then.