Create shared view for word groups.
The rhyme results and synonym results used the exact same layout. This just collapses the View portion into a shared module. This is my first shared module within a view folder. Right now, I'm not doing anything special (ie, no starting with an `_` or placing in sub-folder). We'll see if I refactor this later.