Add share-viewing.
When a person views a share link, I'm going to record the viewing so that the poem author can have some insight into whether or not a given poem has been seen. I'm using htmx to try and post the recording asynchronously, after the share link has been loaded. This is an attempt to prevent social-graph and other bots from being recorded as viewership (which would happen if poem share link was posted to a public forum, such as LinkedIn or Facebook).

This isn't complete - some `todo` elements were left in for now.