Don't log author viewing as share viewing.
Since an author may want to see what the share experience or a given poem is like, we probably don't want to actually log the author's viewing of the share link. Of course, this is only true if the author is viewing from their own browser with an authenticate session. If the author is viewing a link that they send via SMS (for example), the link that opens up in their mobile browser won't be opening up to an authenticated session and therefore will be seen as a loggable viewing.