Create Alpine.js auto-submit directive.
This is intended to be applied on the Submit button that will auto-submit the form. This will give parent directives (ex, `x-prevent-double-submit`) a change to initialize before the form is submitted.

I'm trying to create granular directives as championed by the Basecamp / Hotwire team. I don't think it makes sense all the time - some logic should be in name-spaced controllers. But, some features can be global.