Removing commented-out form token.
When I started this app, I was going to be submitting one-time-use form-tokens with critical mutations (such as revoking share links). But, I wasn't able to justify it to myself. Meaning, what's the meaningful difference between "deleting a poem" and "clearing the poem content". And, if I can't create a meaningful distinction in my mind, then I might as well remove it for now. I can always bring it back for things like "deleting an account" (though that would likely entail an email confirmation).