Restart the ColdFusion application after datasource failure.
Every now and then, the per-application datasource just fails to be created when the ColdFusion application starts for the first time. This adds some special logic to the error handling to look for the datasource failure specifically; and, when detected, shuts down the application via `applicationStop()`. The next request should restart the application and the datasource will almost certainly work.