Allow playground poem to be imported.
All this really does is signal (in the URL) that the "Add Poem" page is being accessed from the playground; and, when true, it will prepopulate the `content` field with the `LocalStorage`-persisted playground content.

This process leaves the playground poem in place (within the `LocalStorage`). I figured this was safer than trying to remove it.