Add a `Go` directory for short-links.
This is normally something I would do with URL rewriting (essentially catching 404-requests and then pattern-matching them against a set of possible values). In this project, I'm trying to keep things as simple as possible. Instead of using URL rewriting, I'm literally just creating folders that represent the "stubs" / "slugs" for the short-links. This is more work, but also something that I can easily commit to the repository (ie, it's not some IIS rule in production or Box configuration in development).

I might very well live to regret it :P