Show partial viewing data for share link.
Now that I'm using Cloudflare's geolocation headers to record the Country / Region / City for a given request, I can start to expose some of that information to the poem author. Obviously, there's a balance in security and privacy here. I'm not going to reveal too much information (ex, I don't output the request IP address); but, I am showing the _broad location_. This experience is less so to expose the share link consumer, and more so to let the poem author know if / when the share link was viewed.

This may prove to be useless in the long run; but, I like it for now.