Provide poem editor-to-composer pathway.
The basic editor for poems is just plain-text. This update provides a way to jump from the plain-text editor over to the composer. In the long run, I'm not 100% convinced that I need two different experiences. But, dropping people right into the composer might be both overwhelming and not what people want, necessarily. So, for the time being, I'm just going to provide a cross-submission.