Un-disable buttons on back-button navigation.
The `x-prevent-double-submit` Alpine directive is disabling buttons in order to prevent accidental double-submissions. But, this was leaving the buttons disabled when the user hit the back-button. Now, I'm listening for the `pageshow` event in order to put the form back into a submittable state. There are probably strange edge-cases where I would _not_ want this to happen; but, for the moment, I can't think of any that would be largely relevant.