Accommodate Datamuse request timeout.
By default, the request timeout for the application is 5-seconds. Then, any longer-running pages need to explicitly set a reasonable timeout, as needed. In the poem composer, rhymes, synonyms, and syllable requests are all ultimately fulfilled by an underlying HTTP call to the Datamuse API; which uses a `cfhttp` call with a 5-second timeout. Which means, if it errors, there may not be enough time for the error to be propagated down to the user.

To help propagate errors without causing a page request timeout error, I'm increasing the request timeout on the few pages that use the `WordService.cfc` (which is using the `DatamuseGateway` under the hood). To this end, I'm "asking" the `WordService` what it's max-timeout might be (since the "Datamuse" angle is an implementation detail).

I don't love the tight coupling; but, I'm not sure what else to do about it.