Locking down `*.{js,less}` to a View File
When I slerp-in all the `.js` and `.less` files associated with the client builds, I'm going to mandate that they are only on files that end with `.view.`. This is just an attempt to enforce some sort of soft standard across the app. And, makes it a little more clear when reading the ingress files how everything fits together.