Use auth-context as implied context on cross-cutting views.
The "Access" models are meant to lock-down access to a given model; or, list of models within a given parent context. But, in some views, we have neither of those. For example, when looking at a list of poems or when looking at all share links. Both of those views are, at most, in the context of the authenticated user. I don't yet have a great pattern for dealing with this elegantly. So, for now, I'm just going to use the `authContext` as the implied context for these cross-cutting views.