Auto-submit composer forms on `input` event.
I'm not 100% sold that this is a good idea; but, I'm finding that I often want to have the composer rhymes and synonyms form auto-submit when I change any of the non-text fields. Usually this is for the radio-button changes. I'm now wiring into those changes via a dedicated CSS class hook.