Move cross-cutting list of shares to `global` folder.
I didn't like having a `share` and a `shareAll` folder right next to each other. I'm moving the "shares across all poems" into a `poem/global` folder. I want the default layout of the folders to be more aligned with "click paths"; and then the `global` folder can represent anything that's cross-cutting (but still in the bounds of a conceptual resource, such as "poems"). Right now it's just shares across all poems; but it could be deleting all shares across all poems, or all viewings, etc.