Trying to get UTC to work properly.
Embarrassed to say that I'm having a bear of time getting all the date/times to work when going through the persistence mechanism. I've finally broken down and just set the server's timezone to be UTC. I had _hoped_ to not have to worry about the timezone and just handle all the conversions in the code. But, I'm having issues with the JDBC driver / Connector/J, which seems to really want to use the server's timezone when converting.