Adding the service layer for poem tagging.
This builds on the model layer, adding orchestrated CRUD operations for Tags as well as associating them with poems. There's still no UI for it;  but that will come next.

I'm still not loving the overall flow of data in some cases. I think I need to start making more data _optional_. I started playing around with that little in the `updatePoem()` method. I also don't know why I've prefixed most args with `poem***`, but not the `userID`. I think that's a hold-over from when I was thinking about passing the `userID` in as the author; but now that this in the `authContext`, I think that `userID` should become `poemUserID` in the service-layer.