Allow tags to be selected on poem.
This allows a 1:N association between tags and poems. I'm not sure that I love the term "tag" now that I have it in place — I feel like really what I am trying to do is create a "collection" of poems. One reason that I didn't want to use the term collection is because it's a native tag in ColdFusion, and I just find the auto-complete in the IDE to be annowing. But, I might re-think that.

There's still no colorful rendering of the tag.