Make service arguments uniformly named.
I was passing `userID` as a non-model-prefixed value, but the rest of the service arguments were all prefixed with the model name (ex, `poemTagID`).

That said, now that I have this in place, I'm not sure that I truly need the model prefixing. I think I used to need it to differentiate between the `userID` of the authenticated user and the `userID` being associated with a specific model. However, now that I'm passing in a uniquely-named `authContext`, I don't think I'll run into this issue very often at all. I might remove the model prefixing. Then, if I ever need to pass-in multiple model arguments, I can also prefix _as needed_, not as a rule.