Moving the poem tagID to the collectionID.
This is a one-time migration to move the gesture of the previous tagging to the notion of the current collection.

I think semantically, it probably makes to be able to associate a poem with more than one collection; but for the sake of simplicity of UI, I'm keeping it a one-to-one association until it proves to be a limitation.