Making "maybe" more of a first class citizen.
I've been using the notion of `maybe` for a while; but it's just been a common pattern more than a first class data structure. This commit tries to evolve the notion into something more structures. Now, there are `maybeNew()` and `maybeSetValue()` (and `maybe.setValue()`) global methods to make this more usable.