Remove warning message about docker pull.
Docker was attempting to pull images for my services; but my images only ever exist (and are built) locally. Adding the `pull_policy` prevents Docker from every tryting to pull. At least I think - the documentation is a little confusing:

https://docs.docker.com/reference/compose-file/services/#pull_policy

> **`pull_policy`** defines the decisions Compose makes when it starts to pull images. Possible values are:
>
> `never`: Compose doesn't pull the image from a registry and relies on the platform cached image. If there is no cached image, a failure is reported.

I _assume_ what they mean is that a "failure is reported" only if the image can't be **built** first. But, I'm not 100% sure.