Replace hard-coded for/id attributes with UI utility.
This is an attempt to simplify some of the label `for` and input `id` attributes; as well as the `aria-describedby` attributes. There were a few in the composer view(s) that I couldn't change because the `id` value was also being consumed by HTMX elsewhere on the page and I need the predictability. But, I can probably do some refactoring there in the future.