Initial pass on Open Graph image for poem share.
This is a work-in-progress for the social platform Open Graph image generation for poem share links. I just want to get _something_ in place so that I can reset my thinking and start to consider how I actually want to move this code around and how the caching will work.