Adding logging to Open Graph image generation.
Since I can't really test the functionality of the Cloudflare CDN locally, I'm going to include some logging so that I can see when images are generated in production. Then, I can hit that end-point (or try to hit that end-point) and see if the caching is working as I would hope.