Include image "version" in Open Graph URL.
The version allows the URL to change as the contents of the poem / author change. And, since we're validating the version prior to image generation, it should allow us to lock-down the URL and enforce caching so that we don't allow the CPU to be saturated by a malicious actor.