Minor refactor for scratch disk.
I'm trying to make the use of "scratch space" just a little bit more formal in the application.

As an aside, I like to use an explicit scratch space instead of the built-in `getTempDirectory()` because it will give me more insight if things go wrong. Meaning, since the scratch directory is local to the application, I'll have an opportunity to see orphaned files and folders if they don't get deleted automatically. In contrast, I have no idea where the `getTempDirectory()` even is; so I'll never see it if files start to accrue there for reasons that need to be investigated.