Allow `f` key to focus first input.
I'm hesitant to use the `autofocus` attribute as I've read that is has several accessibility (a11y) concerns. But, I still like the idea of being able to quickly move focus to the form. I've decided to wire the `f` key as the focus trigger. It matches "f" for both "Form" and "Focus", so it just kind of works. Plus, it's on the left side of the keyboard, away from the mouse (for the majority of people), which means that it's easy for most people to reach.