Move `x-keyed-focus` to input, not form.
As I was working through my blog post about the `x-keyed-focus` directive, it became more clear to me that the smarter move was to put the directive on the target input, not on the form. Putting it on the form does create some more flexibility; but, for what end? It also brings more confusion. By moving the directive to the input itself, I get to remove a bunch of the internal logic.

Read post: https://bennadel.com/4849