Simplifying Application.cfc logic.
I had mappings and error handling logic in place that didn't really need to be there. I'm simplifying the logic as much as I can. In an application with this architecture, all runtime / user-facing errors are handled in the main `try`/`catch` block. As such, any errors that bubble-up to the `Application.cfc` only need minimal treatment. If an error makes it that far, the application is in a bad state and attempting to recover in any clever way is just ill-informed.