Adding decode-mappings to gateway constructor.
I'm not 100% sure on this one; but, I'm adding the `decodeMappings` to the gateway constructor. This way, the gateways can just call `decodeColumns()` without having to pass-in the structure. It makes the code a little cleaner looking. But, I don't love that it's asymmetric. Meaning, all the _encoding_ of columns is still inline. I wonder if I should have an `encodeColumns()` call. But, that might start to get too abstract. Baby steps.