Baby steps to responsive tables.
Responsive tables are not easy or obvious. As a first baby step, I'm just wrapping the tables in a horizontally-scrollable DIV. This means that I can keep the basic layout of the table in-tact without completely destroying the UI of the mobile experience. It's not a great approach, but it's basically no effort. And it will at least get me to an MVP of a responsive experience.