Add "copy link" buttons for public share links.
This add a `Copy Link` button next to the places in which the public URL is linked in the UI.

I'm starting to feel like I don't have a cohesive sense of how the JavaScript parts of this application fit together. I think part of it is that I have two semi-competing frameworks (htmx and Alpine), which both provide some overlapping JavaScript functionality. But, the other part is that since I have no holistic architecture, I'm somewhat discovering patterns as I go. I **do like** that it's rather simple to apply JavaScript and LessCSS to a given view; but, I don't have a good instinct for how to share JavaScript utilities across the app.

This came to a point today because I wanted to get at the `$dispatch()` magic in Alpine. I eventually figured out how to hack it into place by using the `$data()` method. But, really, if I just had a centralized **utils**, I could have my own `dispatch()` utility. But, I'm not sure how that would work exactly. Something to noodle on.