Show log count in local development widget.
Logs are often representative of something going wrong. When it's an "Unprocessable Entity" issue, it's obvious because I'm the one causing it. But, if it's due to a background task / API call, it won't be obvious. I want to be always-aware of any logs that need to be reviewed and addressed prior to deployment.