Move from `public` to `static`.
I was using `public` for my assets as some strange hold-over from the Node.js/Express.js world (in which the Node runtime has to create both the "application" and the "web"). With ColdFusion, especially running on IIS, the web server is created for you; and often the `wwwroot` folder _is_ the "public" folder. As such, to have a public folder named "public" is both redundant and misleading (since other folders are also public). As such, I'm renaming this folder to be `static` since it's for non-CFML/non-dynamic assets that get linked. Yes, the contents of the folder are changed during _build time_; but, at _run time_ they are static.