Move the "go" folder into the routing.
I didn't love having part of the application existing outside of the `/client` folder. It created confusion around error handling, asset building, and the general consistency in the application architecture. I've shifted to moving the `go` logic _into_ the `/client` directory; and then using the physical "go/" folder as nothing more than a `location()`-based redirect into the normal application routing. So, it's kind of just like an alias now.

It does add a little more complexity; but, the up-side is that I get the full build-system now - including a special ingress for the `go` overview itself.