Add IP-location to user session tracking.
Now that I know that Cloudflare can give me the IP-based city, region, and country code, I think it makes sense to start tracking that data in the user sessions. I can still link to the IPInfo site; but, why not just have that information right in the list so that the user doesn't have to wonder.

I'm also removing the only `Partial.cfc` that I had in the app - I think I can make due with just the `ListGateway.cfc`, and do any needed massaging of data in the controller itself. I'm trying to simplify wherever I can.